Pipe: Reuse clusters across related integration tests
